git-svn-id: http://chianti.ucsd.edu/svn@15532 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ csplugins/trunk/pasteur/mathieu/Tunables/src/Command/input.java
@@ -28,7 +28,7 @@ public class input<O extends Comparable<String>> implements command {
 	public Group Group2;
 	
 	@SuppressWarnings("unchecked")
-	@Tunable(description="BoundedObject",type=Bounded.class,flag=Param.Integer)
+	@Tunable(description="BoundedObject",type=Bounded.class,flag=Param.Double)
 	public Bounded<O> boundObject= new Bounded<O>((O)"0",(O)"100", true, true);
 	
 	@Tunable(description="Group",type=Group.class,flag=Param.Nothing)
